Allow optional explicit delimeter
The add method now accepts input_strings that
include an excplicitly declared delimeter on the
first line of the string.

When such a delimiter is provided, it becomes the
only accepted "default" delimeter and commas are
no longer used. When it is not provided, commas
and new lines are accepted and treated as the
default delimeter